Move common components to datasource.rs, will be utilized for parquet (+ other) datasources / writers. Primary types and polars as csv translation medium
