+++ src/gcp_bigquery.rs
@@ -1,10 +1,9 @@
 use crate::{
     csv::CsvWriter,
-    datasource::DatasourceWritable,
+    datasource::{load_table_configs, read_csv_to_polars, DatasourceWritable},
     types::{IndexerContractMapping, IndexerGcpBigQueryConfig},
 };
 use async_trait::async_trait;
-use csv::ReaderBuilder;
 use gcp_bigquery_client::{
     error::BQError,
     model::{
@@ -14,23 +13,10 @@ use gcp_bigquery_client::{
     Client,
 };
 use indexmap::IndexMap;
-use phf::phf_ordered_map;
+// use phf::phf_ordered_map;
 use polars::prelude::*;
 use serde_json::Value;
-use std::{any::Any, collections::HashMap, fs::File};
-
-///
-/// Columns common to all tables
-/// Each table's columns will include the following commmon fields
-///
-static COMMON_COLUMNS: phf::OrderedMap<&'static str, &'static str> = phf_ordered_map! {
-    "indexed_id" => "string",  // will need to generate uuid in rust; postgres allows for autogenerate
-    "contract_address" => "string",
-    "tx_hash" => "string",
-    "block_number" => "int",
-    "block_hash" => "string",
-    "timestamp" => "int"
-};
+use std::{any::Any, collections::HashMap};
 
 /// Query client
 /// Impl for this struct is further below
@@ -78,71 +64,6 @@ pub async fn init_gcp_bigquery_db(
     }
 
     Ok(gcp_bigquery)
-}
-
-///
-/// This function merges the common columns specified in COMMON_COLUMNS with the
-/// columns specified in the reth-indexer-config file for the particular event type to parse
-/// The types are ordered to allow for consistent column order within the GCP tables
-/// This is done for each table specified in the indexer
-///
-/// # Arguments
-///
-/// * `indexer_event_mappings` - list of all event mappings
-pub fn load_table_configs(
-    indexer_event_mappings: &[IndexerContractMapping],
-) -> HashMap<String, IndexMap<String, String>> {
-    //  Config map for all tables
-    //  Ordered by common columns, then by configured mapped
-    let mut all_tables: HashMap<String, IndexMap<String, String>> = HashMap::new();
-
-    for mapping in indexer_event_mappings {
-        for abi_item in mapping.decode_abi_items.iter() {
-            let table_name = abi_item.name.to_lowercase();
-            let column_type_map: IndexMap<String, String> = abi_item
-                .inputs
-                .iter()
-                .map(|input| {
-                    (
-                        input.name.clone(),
-                        solidity_type_to_db_type(input.type_.clone().as_str()).to_string(),
-                    )
-                })
-                .collect();
-
-            let common_column_type_map: IndexMap<String, String> = COMMON_COLUMNS
-                .into_iter()
-                .map(|it| (it.0.to_string(), it.1.to_string()))
-                .collect();
-
-            let merged_column_types: IndexMap<String, String> = common_column_type_map
-                .into_iter()
-                .chain(column_type_map)
-                .collect();
-
-            all_tables.insert(table_name, merged_column_types);
-        }
-    }
-
-    all_tables
-}
-
-///
-/// Maps solidity types (in indexer config) to bigquery column types
-/// The configuration / reth uses solidity types, we map these to types that we'll use in GCP
-/// bigquery
-///
-/// # Arguments
-///
-/// * `abi_type` - the ABI type, specified as a string
-fn solidity_type_to_db_type(abi_type: &str) -> &str {
-    match abi_type {
-        "address" => "string",
-        "bool" | "bytes" | "string" | "int256" | "uint256" => "string",
-        "uint8" | "uint16" | "uint32" | "uint64" | "uint128" | "int8" | "int16" | "int32"
-        | "int64" | "int128" => "int",
-        _ => panic!("Unsupported type {}", abi_type),
-    }
 }
 
 impl GcpBigQueryClient {
@@ -304,48 +225,12 @@ impl GcpBigQueryClient {
     ///  * `table_name` - name of table for dataset
     ///  * `csv_writer` - csv writer reference
     pub async fn write_csv_to_storage(&self, table_name: &str, csv_writer: &CsvWriter) {
-        let dataframe = self.read_csv_contents(table_name, csv_writer.path());
+        let column_map = &self.table_map[table_name];
+        let dataframe = read_csv_to_polars(csv_writer.path(), column_map);
         let bq_rowmap_vector =
             self.build_bigquery_rowmap_vector(&dataframe, &self.table_map[table_name]);
         self.write_rowmaps_to_gcp(table_name, &bq_rowmap_vector)
             .await;
-    }
-
-    ///
-    /// Read CSV
-    /// Return dataframe, transform to vector and write to GCP
-    ///
-    /// # Arguments
-    ///
-    /// * `table_name` - name of table for dataset
-    /// * `path` - path to csv file
-    pub fn read_csv_contents(&self, table_name: &str, path: &str) -> DataFrame {
-        //  Get list of columns in order, from csv file
-        let file = File::open(path).unwrap();
-        let mut rdr = ReaderBuilder::new().has_headers(true).from_reader(file);
-        let headers = rdr.headers().unwrap();
-        let column_names: Vec<String> = headers.iter().map(String::from).collect();
-
-        //  Build column data types for polars dataframe, from column mapping
-        let column_map = &self.table_map[table_name];
-        let plr_col_types = Some(Arc::new(
-            column_names
-                .iter()
-                .map(|name| match column_map[name].as_str() {
-                    "int" => Field::new(name, DataType::Int64),
-                    "string" => Field::new(name, DataType::Utf8),
-                    _ => panic!("incompatible type found"),
-                })
-                .collect(),
-        ));
-
-        //  Read polars dataframe, w/ specified schema
-        CsvReader::from_path(path)
-            .expect("file does not exist")
-            .has_header(true)
-            .with_schema(plr_col_types)
-            .finish()
-            .unwrap()
     }
 
     ///
